Overhaul of the Vertex Animation Texture
This is a major over haul of how the vertex animation texture is designed. It centrifies the backbone from four unqiue exporters to a common exporter. It adds additional functionality and standardizes the shaders and the export attributes across all the exported files. It even will update the Unity files.